requirements-simple.txt
supabase>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
cryptography>=40.0.0
pydantic>=1.10.0
python-dateutil>=2.8.0
